Occi: use Guice for DI
Guice used to avoid all these nasty singletons and static fields!
The OcciModule class is responsible for the "runtime wiring" of things.

Added a simple inmemory DB implementation for tests
One test for OrientDB will still be failing as we need to find a better way to fix it.

Resources are now stored and read from DB instead of keeping a static map + DB.